% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{redefine weights assigned to models in an ensemble}
\usage{
calibrate(ensemble, data, label)
}
\arguments{
\item{ensemble}{object of class 'ml_ensemble'}

\item{data}{calibration data, e.g. matrix or data frame as
appropriate for the model ensemble}

\item{label}{vector with expected predictions, e.g. numeric values for
regression, or integer/character values for classification}
}
\value{
object of class ml_ensemble
}
\description{
redefine weights assigned to models in an ensemble
}
\examples{

}
