% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_one}
\alias{calibrate_one}
\title{build one calibration model}
\usage{
calibrate_one(raw, label, j, family)
}
\arguments{
\item{raw}{list of matrices with predictions}

\item{label}{matrix with expected predictions}

\item{j}{integer, column to consider}

\item{family}{description of error distribution}
}
\value{
glm model
}
\description{
build one calibration model
}
