% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{cut_to_features}
\alias{cut_to_features}
\title{cut a matrix, data.frame, or data table to certains columns}
\usage{
cut_to_features(x, col_names)
}
\arguments{
\item{x}{matrix, data.frame, or data.table}

\item{col_names}{character, vector with feature names to keep}
}
\value{
object similar to x, cut to certain columns only
}
\description{
cut a matrix, data.frame, or data table to certains columns
}
