% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{label_matrix}
\alias{label_matrix}
\title{standardize labels into matrix form}
\usage{
label_matrix(label, one_hot)
}
\arguments{
\item{label}{vector with values}

\item{one_hot}{logical, determines if label matrix has one column
with values, or a matrix using one-hot encoding}
}
\value{
matrix with one row per prediction, one column per label class
}
\description{
standardize labels into matrix form
}
