% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_ensemble.R
\name{ml_ensemble}
\alias{ml_ensemble}
\title{define an ensemble of ml_model objects}
\usage{
ml_ensemble(name = NULL, hooks = list(), description = NA)
}
\arguments{
\item{name}{character, a unique name for the model}

\item{hooks}{object of class ml_hooks, or a list of ml_hook objects}

\item{description}{character, used in construction of plain-language
summaries of the ensemble; recommended to consist of a single sentence
fragment starting with a verb, e.g. 'integrates predictions'.}
}
\value{
object of class ml_ensemble
}
\description{
define an ensemble of ml_model objects
}
\examples{

#empty ensemble
ensemble = ml_ensemble()

# ensemble with several models
lm_1 = lm(y~x, data=data.frame(x=1:2, y=-0.2 + 1:2))
lm_2 = lm(y~x, data=data.frame(x=1:2, y=+0.2 + 1:2))
ensemble = ml_model(lm_1) + ml_model(lm_2)

}
