% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_hook.R
\name{ml_hook}
\alias{ml_hook}
\title{define one hook for a ml_model object}
\usage{
ml_hook(f, name = NULL, type = c("pre", "post"), order = 1)
}
\arguments{
\item{f}{function}

\item{name}{character, a name/identifier for the hook}

\item{type}{character, vector of feature names. If NULL,
feature names will be guessed from the model object}

\item{order}{numeric, determines the order/priority of a hook when it is
used within a collection of many other hooks}
}
\value{
the same object with a new class 'ml_model'
}
\description{
define one hook for a ml_model object
}
\examples{

# define a hook function that replaces NA of "x" in a data frame
fill_NA <- function(d) {
  d[is.na(d$x), "x"] <- mean(d$x, na.rm=TRUE)
  d
}
# define a hook using fill_NA as a pre-processing operation
hook <- ml_hook(fill_NA, type="pre")
hook

}
