% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_hooks.R
\name{ml_hooks}
\alias{ml_hooks}
\title{define a list of ml_hook objects}
\usage{
ml_hooks(x = list())
}
\arguments{
\item{x}{list of ml_hook objects}
}
\value{
a list-like object
}
\description{
define a list of ml_hook objects
}
\examples{

# empty collection of hooks
my_hooks <- ml_hooks()
my_hooks

# prep: hook functions
hook_fun_1 <- function(x) { x**2 }
hook_fun_2 <- function(x) { sqrt(2) }
# list of hook objects, and object containing the list of hooks
hook_list <- list(ml_hook(hook_fun_1), ml_hook(hook_fun_2))
my_hooks <- ml_hooks(hook_list)
my_hooks

}
