% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{ml_model}
\alias{ml_model}
\title{define an object as an ml_model type}
\usage{
ml_model(
  model,
  name = NULL,
  feature_names = NULL,
  label_names = NULL,
  hooks = list()
)
}
\arguments{
\item{model}{object with a machine learning model (e.g. lm, glm, xgboost)}

\item{name}{character, a unique name for the model}

\item{feature_names}{character, vector of feature names. If NULL,
feature names will be guessed from the model object}

\item{label_names}{character, vector of labels for multi-class
classification.}

\item{hooks}{objects of class ml_hooks, or a list of ml_hook objects}
}
\value{
an object holding the model and book-keeping metadata
}
\description{
define an object as an ml_model type
}
\examples{

lm_1 = lm(y~x, data=data.frame(x=1:2, y=1:2))

# default constructor
ml_model(lm_1)

# use a custom name
ml_model(lm_1, name="linear_model_1")

}
