% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_ensemble.R
\name{+.ml_model}
\alias{+.ml_model}
\title{create an ensemble from two models}
\usage{
\method{+}{ml_model}(m1, m2)
}
\arguments{
\item{m1}{model object}

\item{m2}{model object}
}
\value{
a model ensemble

lm_1 = lm(y~x, data=data.frame(x=1:2, y=-0.2 + 1:2))
lm_2 = lm(Y~x, data=data.frame(x=1:2, y=+0.2 + 1:2))
ensemble = ml_model(lm_1) + ml_model(lm_2)
}
\description{
create an ensemble from two models
}
