% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ml_ensemble}
\alias{predict.ml_ensemble}
\title{use an ensemble of models (ml_ensembl) to predict output for a new dataset}
\usage{
\method{predict}{ml_ensemble}(object, data, type = "response", ...)
}
\arguments{
\item{object}{object of class ml_ensemble}

\item{data}{object with data to operate on, e.g. matrix or data frame as
appropriate for the ensemble of models}

\item{type}{character, passed to predict()}

\item{...}{other arguments, passed to predict()}
}
\value{
object of class ml_ensemble
}
\description{
use an ensemble of models (ml_ensembl) to predict output for a new dataset
}
