% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ml_model}
\alias{predict.ml_model}
\title{use an ml_model to predict output for a new dataset}
\usage{
\method{predict}{ml_model}(object, data, type = "response", ...)
}
\arguments{
\item{object}{object of class ml_model}

\item{data}{object with data to operate on, e.g. matrix or data frame}

\item{type}{character, code passed to predict()}

\item{...}{other arguments passed to predict()}
}
\value{
prediction of the model "object" on a dataset "data"
}
\description{
use an ml_model to predict output for a new dataset
}
