% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prepare_newdata}
\alias{prepare_newdata}
\title{prepare a data object so that it includes all required features}
\usage{
prepare_newdata(newdata, feature_names)
}
\arguments{
\item{newdata}{object with data to operate on, e.g. matrix or data.frame}

\item{feature_names}{character, vector with feature names}
}
\value{
object with data holding expected features; some features in newdata
may be omitted, and other features added with NA values
}
\description{
prepare a data object so that it includes all required features
}
