% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ml_ensemble}
\alias{print.ml_ensemble}
\title{Display a simple summary of an ml_ensemble object}
\usage{
\method{print}{ml_ensemble}(x, ...)
}
\arguments{
\item{x}{ml_ensemble}

\item{...}{other parameters (not used)}
}
\description{
Display a simple summary of an ml_ensemble object
}
\examples{

lm_1 = lm(y~x, data=data.frame(x=1:2, y=-0.2 + 1:2))
lm_2 = lm(Y~x, data=data.frame(x=1:2, y=+0.2 + 1:2))
ensemble = ml_model(lm_1) + ml_model(lm_2)
print(ensemble)

}
