% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ml_hook}
\alias{print.ml_hook}
\title{Display a simple summary of an ml_hook object}
\usage{
\method{print}{ml_hook}(x, ...)
}
\arguments{
\item{x}{object of class ml_hook}

\item{...}{other parameters (not used)}
}
\description{
Display a simple summary of an ml_hook object
}
\examples{

# hook function for normalizing matrices so that rows sum to unity
my_fun = function(x) {
  x / apply(x, 1, sum)
}
my_hook = ml_hook(my_fun)
my_hook

}
