% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ml_hooks}
\alias{print.ml_hooks}
\title{Display a simple summary of a list of ml_hook objects}
\usage{
\method{print}{ml_hooks}(x, ...)
}
\arguments{
\item{x}{object of class ml_hooks}

\item{...}{other parameters (not used)}
}
\description{
Display a simple summary of a list of ml_hook objects
}
\examples{

my_fun = function(x) {
  x / apply(x, 1, sum)
}
my_hooks = ml_hooks() + ml_hook(my_fun)
my_hooks

}
