% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ml_model}
\alias{print.ml_model}
\title{Display a simple summary of an ml_model object}
\usage{
\method{print}{ml_model}(x, ...)
}
\arguments{
\item{x}{ml_model}

\item{...}{other parameters (not used)}
}
\description{
Display a simple summary of an ml_model object
}
\examples{

lm_1 = lm(y~x, data=data.frame(x=1:2, y=1:2))
model = ml_model(lm_1)
print(model)

}
