% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{raw_predict_ensemble}
\alias{raw_predict_ensemble}
\title{use an ensemble of models to predict on new data}
\usage{
raw_predict_ensemble(ensemble, data, ...)
}
\arguments{
\item{ensemble}{object of class ml_ensemble}

\item{data}{object with data}

\item{...}{other arguments passed to predict}
}
\value{
list of raw predictions
}
\description{
use an ensemble of models to predict on new data
}
