% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{standardize_raw}
\alias{standardize_raw}
\title{standardize predictions into matrix form}
\usage{
standardize_raw(raw)
}
\arguments{
\item{raw}{list of predictions as vectors or matrices}
}
\value{
list of predictions, always as matrices
}
\description{
standardize predictions into matrix form
}
