% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ml_ensemble}
\alias{summary.ml_ensemble}
\title{Display a summary of an ml_ensemble object}
\usage{
\method{summary}{ml_ensemble}(object, ...)
}
\arguments{
\item{object}{ml_ensemble}

\item{...}{other parameters (not used)}
}
\description{
Display a summary of an ml_ensemble object
}
\examples{

lm_1 = lm(y~x, data=data.frame(x=1:2, y=-0.2 + 1:2))
lm_2 = lm(y~x, data=data.frame(x=1:2, y=+0.2 + 1:2))
ensemble = ml_model(lm_1) + ml_model(lm_2)
summary(ensemble)

}
