% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ml_model}
\alias{summary.ml_model}
\title{Display a description of an ml_model object}
\usage{
\method{summary}{ml_model}(object, ...)
}
\arguments{
\item{object}{ml_model}

\item{...}{other parameters (not used)}
}
\description{
Display a description of an ml_model object
}
\examples{

lm_1 = lm(y~x, data=data.frame(x=1:2, y=1:2))
model = ml_model(lm_1)
summary(model)

}
