% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{trim_model_name}
\alias{trim_model_name}
\title{trim a model name into alphanumeric, underscore, and dot}
\usage{
trim_model_name(x)
}
\arguments{
\item{x}{character}
}
\value{
character, model name, suitable as a data frame column
}
\description{
trim a model name into alphanumeric, underscore, and dot
}
