% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_ensemble.R
\name{ml_ensemble}
\alias{ml_ensemble}
\title{define an ensemble of ml_model objects}
\usage{
ml_ensemble(model_name = NULL)
}
\arguments{
\item{model_name}{character, a name}
}
\value{
object of class ml_ensemble
}
\description{
define an ensemble of ml_model objects
}
\examples{

#empty ensemble
ensemble = ml_ensemble()

# ensemble with several models
lm_1 = lm(y~x, data=data.frame(x=1:2, y=-0.2 + 1:2))
lm_2 = lm(Y~x, data=data.frame(x=1:2, y=+0.2 + 1:2))
ensemble = ml_model(lm_1) + ml_model(lm_2)

}
