% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_hook.R
\name{ml_hook}
\alias{ml_hook}
\title{define one hook for a ml_model object}
\usage{
ml_hook(f, name = NULL, type = c("pre", "post"), order = 1)
}
\arguments{
\item{f}{function}

\item{hook_name}{character, a name/identifier for the hook}

\item{hook_type}{character, vector of feature names. If NULL,
feature names will be guessed from the model object}
}
\value{
the same object with a new class 'ml_model'
}
\description{
define one hook for a ml_model object
}
\examples{

lm_1 = lm(y~x, data=data.frame(x=1:2, y=1:2))

# default constructor
ml_model(lm_1)

# use a custom name
ml_model(lm_1, model_name="linear_model_1")

}
