% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_hooks.R
\name{ml_hooks}
\alias{ml_hooks}
\title{define a list of ml_hook objects}
\usage{
ml_hooks(x = list())
}
\arguments{
\item{x}{list of ml_hook objects}
}
\value{
a list-like object
}
\description{
define a list of ml_hook objects
}
\examples{

my_hooks <- ml_hooks()
my_hooks

}
