% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model.R
\name{ml_model}
\alias{ml_model}
\title{define an object as an ml_model type}
\usage{
ml_model(model, model_name = NULL, feature_names = NULL, label_names = NULL)
}
\arguments{
\item{model}{object with a machine learning model (e.g. lm, glm, xgboost)}

\item{model_name}{character}

\item{feature_names}{character, vector of feature names. If NULL,
feature names will be guessed from the model object}

\item{label_names}{character, vector of labels for multi-class
classification.}
}
\value{
the same object with a new class 'ml_model'
}
\description{
define an object as an ml_model type
}
\examples{

lm_1 = lm(y~x, data=data.frame(x=1:2, y=1:2))

# default constructor
ml_model(lm_1)

# use a custom name
ml_model(lm_1, model_name="linear_model_1")

}
